- added possibility to play animations with different speed - enemy level can be a scalar - fix king tomato low life now is higher - fix garlic particles are burst not fountain

+++ trunk/MyGame/MyGame/code/Gameplay/Enemies/Enemy.cs
@@ -8,9 +8,9 @@ namespace MyGame
 {
     public class Enemy : CollidableEntity2D
     {
-        public int enemyLevel { get; set; }
+        public float enemyLevel { get; set; }
 
-        public Enemy(string entityName, Vector3 position, float orientation, int enemyLevel, int id = -1)
+        public Enemy(string entityName, Vector3 position, float orientation, float enemyLevel, int id = -1)
             : base("enemies", entityName, position, orientation, Color.White, id)
         {
             this.enemyLevel = enemyLevel;
